+++ packages/expo-image/android/src/main/java/expo/modules/image/ExpoImageModule.kt
@@ -162,6 +162,10 @@ class ExpoImageModule : Module() {
         view.recyclingKey = recyclingKey
       }
 
+      Prop("disableDownscaling") { view: ExpoImageViewWrapper, disableDownscaling: Boolean? ->
+        view.disableDownscaling = disableDownscaling ?: false
+      }
+
       OnViewDidUpdateProps { view: ExpoImageViewWrapper ->
         view.rerenderIfNeeded()
       }